Make `CMat` public.
This is horrible, and I know it.

The reason to do it is merely to be able to integrate with other C
libraries that use OpenCV structures in their interfaces (for example,
certain parts of dlib).

My options are:
- Repeat code from this library in the integration with said frameworks.
- Make this public.
- Do black magic.

Making this public seems to be less horrible alternative.

Everything is terrible.